7zip

+++ Scripts/7zip-creator.ps1
@@ -1,6 +1,15 @@
 ﻿ 
+
+[CmdletBinding()]
+Param
+    (
+    [string] $ScriptPath = "C:\code\EricCote.github.io\scripts\install-oracle.ps1", 
+    [string] $Destination = $(Join-Path  $env:USERPROFILE "downloads\"),
+    [switch] $Uninstall = $false 
+    )
  
-$dl=$(Join-Path  $env:USERPROFILE "downloads\");
+
+$dl=$Destination
 
 if(-not (test-path "C:\Program Files\7-Zip")) {
 
@@ -15,7 +24,9 @@ if(-not (test-path "C:\Program Files\7-Zip")) {
     "install 7zip..."
     & $fileDest /S /D="c:\program files\7-Zip" | out-null 
 
-    }
+}
+
+
 #    "downloading extras..."
 #    $fileSource = "http://7-zip.org/a/7z1604-extra.7z"
 #    $fileDest= ($dl + "7z1604-extra.7z")
@@ -26,27 +37,63 @@ if(-not (test-path "C:\Program Files\7-Zip")) {
 
 
 
-    "Downloading sfx"
-   
-    $fileSource = "https://github.com/EricCote/AfiSetup/raw/master/7z/7zS2con.sfx"
-    $fileDest= ($dl + "7zS2con.sfx")
 
-    if (-not (test-path $fileDest)) { 
+   
+$fileSource = "https://github.com/EricCote/AfiSetup/raw/master/7z/7zS2con.sfx"
+$fileDest= ($dl + "7zS2con.sfx")
+
+if (-not (test-path $fileDest)) {
+    "Downloading sfx" 
     $wc = new-object System.Net.WebClient ;
     $wc.DownloadFile($fileSource, $fileDest) ;
     $wc.Dispose();  
-    }
+}
+
+   # Originally,I used the sfx from extras. I don't anymore. 
    # "unzipping extras"
    # & "C:\Program Files\7-Zip\7z.exe" l $fileDest
 
+if (-not (Test-Path $ScriptPath)){
+    "script path not there"
+    return "script Path not here!"
 
-    $psFileName="20480.ps1"
+}
 
-    $cmd= @"
-      @%SystemRoot%\sysnative\WindowsPowerShell\v1.0\powershell.exe -ExecutionPolicy bypass -f "%~dp0$psFileName"
+ 
+$scriptName = (dir $ScriptPath)[0].Name
+$scriptBase = (dir $ScriptPath)[0].BaseName
+
+ $cmd= @"
+      @%SystemRoot%\sysnative\WindowsPowerShell\v1.0\powershell.exe -ExecutionPolicy bypass -f "%~dp0$scriptName"
 "@
    
- Set-Content ($dl + "install.cmd") $cmd
+ Set-Content ($dl + "install.cmd") $cmd;
 
- & "C:\Program Files\7-Zip\7z.exe" a  ($dl + "out.7z")  ($dl + '20480.ps1') ($dl + 'install.cmd')  -mx 
- & cmd /c copy /b  ($dl + "7zS2con.sfx") + ($dl + "out.7z")  ($dl + "setup.exe")  
+ & "C:\Program Files\7-Zip\7z.exe" a  ($dl + "out.7z")  ($ScriptPath) ($dl + 'install.cmd')  -mx 
+ & cmd /c copy /b  ($dl + "7zS2con.sfx") + ($dl + "out.7z")  ($dl + $ScriptBase + ".exe")  
+
+
+
+
+    function Get-BasicAuthCreds {
+        param([string]$Username,[string]$Password)
+        $AuthString = "{0}:{1}" -f $Username,$Password
+        $AuthBytes  = [System.Text.Encoding]::Ascii.GetBytes($AuthString)
+        return [Convert]::ToBase64String($AuthBytes)
+
+    }
+
+    $cred=  (Get-Credential -Message 'Enter Github password' -UserName 'eric@coteexpert.com');
+    $BasicCreds = Get-BasicAuthCreds -Username "eric@coteexpert.com" -Password $cred.GetNetworkCredential().password
+   
+    $result = Invoke-RestMethod -Method Post `
+     -Uri "https://uploads.github.com/repos/EricCote/EricCote.github.io/releases/4838718/assets?name=$ScriptBase.exe" `
+     -Credential  $cred `
+     -Headers @{"Authorization"="Basic $BasicCreds"} `
+     -InFile ($dl + $ScriptBase + ".exe") `
+     -ContentType "application/octet-stream" 
+
+
+
+
+
